+++ packages/chain-consumer/src/decomposers/ExchangeProposalDecomposer.ts
@@ -3,7 +3,6 @@ import {
   PerpetualMarketLaunchProposal,
   SpotMarketLaunchProposal,
   SpotMarketParamUpdateProposal,
-  SpotMarketStatusSetProposal,
 } from '@injectivelabs/chain-api/injective/exchange/v1beta1/tx_pb'
 
 export class ExchangeProposalDecomposer {
@@ -13,10 +12,6 @@ export class ExchangeProposalDecomposer {
 
   static spotMarketUpdate(content: Uint8Array) {
     return SpotMarketParamUpdateProposal.deserializeBinary(content)
-  }
-
-  static spotMarketStatusUpdate(content: Uint8Array) {
-    return SpotMarketStatusSetProposal.deserializeBinary(content)
   }
 
   static perpetualMarketLaunch(content: Uint8Array) {