+++ packages/derivatives-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.32](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.31...@injectivelabs/derivatives-consumer@0.0.32) (2021-06-25)
+
+
+### Features
+
+* added batch cancelling for the orders ([239a672](https://github.com/InjectiveLabs/injective-ts/commit/239a6723217301aa86f657d59e9e76721340b1f2))
+
+
+
+
+
 ## [0.0.31](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/derivatives-consumer@0.0.29...@injectivelabs/derivatives-consumer@0.0.31) (2021-06-17)
 
 **Note:** Version bump only for package @injectivelabs/derivatives-consumer